# public/openapi/coreth.yaml
openapi: 3.0.0
info:
  title: Coreth C-Chain API
  description: >-
    JSON-RPC 2.0 API for Avalanche's C-Chain. Since the Helicon upgrade, the C-Chain runs Continuous Execution (ACP-194) and no longer runs Coreth. All methods use POST to the same
    endpoint with the method specified in the request body.
    
    
    **Important Notes on Method Availability:**
    
    • **Public RPC Limitations:** Not all methods are available on public RPC endpoints.
    
    • **Restricted Methods:**
      - `debug_*` methods: NOT available on public RPCs (require node access)
      - `personal_*` methods: removed from every C-Chain node in AvalancheGo v1.15.0  
      - `txpool_*` methods: May be restricted or rate-limited
      - Filter methods (`eth_newFilter`, etc.): May have limited functionality
      - `eth_sign`, `eth_signTransaction` and `eth_sendTransaction`: fail on every C-Chain node, because the node holds no accounts. Sign on the client and use `eth_sendRawTransaction`
    
    • **Full Access:** Run your own node or use a dedicated node service for unrestricted access to all methods.
    
    • **Custom RPC Endpoints:** While this spec provides common public endpoints, most API clients and tools 
    allow you to override the server URL with your own RPC endpoint when making requests.
  version: 1.0.0
servers:
  - url: https://api.avax.network
    description: Avalanche Mainnet (Public RPC)
  - url: https://api.avax-test.network
...
paths:
...
  /ext/bc/C/rpc#eth_getBlockByNumber:
    post:
      summary: eth_getBlockByNumber
      description: Returns the block for a given number. Second param selects full transactions.
      operationId: eth_getBlockByNumber
      tags:
        - eth
      requestBody:
        required: true
        description: JSON-RPC request
        content:
          application/json:
            schema:
              type: object
              properties:
                jsonrpc:
                  type: string
                  example: '2.0'
                method:
                  type: string
                  example: eth_getBlockByNumber
                params:
                  type: array
                  items:
                    oneOf:
                      - type: string
                        description: "Block parameter - 'latest', 'earliest', 'pending', or block number in hex"
                      - type: boolean
                        description: "If true, returns the full transaction objects; if false, only the hashes"
                  minItems: 2
                  maxItems: 2
                  description: "[blockParameter, returnFullTransactions]"
                id:
                  type: integer
                  example: 1
              required:
                - jsonrpc
                - method
                - params
                - id
            examples:
              default:
                value:
                  jsonrpc: '2.0'
                  method: eth_getBlockByNumber
                  params: ['latest', false]
                  id: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONRPCResponse'
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONRPCError'
...
  /ext/bc/C/rpc#eth_getTransactionReceipt:
    post:
      summary: eth_getTransactionReceipt
      description: Returns the receipt of a transaction by transaction hash.
      operationId: eth_getTransactionReceipt
      tags:
        - eth
      requestBody:
        required: true
        description: JSON-RPC request
        content:
          application/json:
            schema:
              type: object
              properties:
                jsonrpc:
                  type: string
                  example: '2.0'
                method:
                  type: string
                  example: eth_getTransactionReceipt
                params:
                  type: array
                  items:
                    type: string
                    description: "Transaction hash (32 bytes, hex encoded with 0x prefix)"
                    pattern: "^0x[a-fA-F0-9]{64}$"
                  minItems: 1
                  maxItems: 1
                  description: "[transactionHash]"
                id:
                  type: integer
                  example: 1
              required:
                - jsonrpc
                - method
                - params
                - id
            examples:
              default:
                value:
                  jsonrpc: '2.0'
                  method: eth_getTransactionReceipt
                  params: ['0xb903239f8543d04b5dc1ba6579132b143087c68db1b2168786408fcbce568238']
                  id: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONRPCResponse'
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONRPCError'
...
components:
  schemas:
...
    JSONRPCResponse:
      type: object
      description: JSON-RPC 2.0 Response
      properties:
        jsonrpc:
          type: string
          example: '2.0'
        result:
          description: The result of the method call
        error:
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
            data:
              description: Additional error data
        id:
          type: integer
          example: 1
      required:
        - jsonrpc
        - id
    JSONRPCError:
      type: object
      description: JSON-RPC 2.0 Error Response
      properties:
        jsonrpc:
          type: string
          example: '2.0'
        error:
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
            data:
              description: Additional error data
          required:
            - code
            - message
        id:
          type: integer
          example: 1
      required:
        - jsonrpc
        - error
        - id